Delete plugins directory
Also:
- Update documentation on mixins
- Add a mixin.py example
- Move ruff config to ruff.toml
- Upgrade to ruff 0.13.1
