--- docs/plugins/index.md
@@ -1,7 +1,0 @@
-# Plugins
-
-<!--intro-start-->
-
-- [cmd2 Plugin Template](plugin_template.md)
-
-<!--intro-end-->
