--- plugins/README.txt
@@ -1,1 +1,0 @@
-For information about creating a cmd2 plugin, see template/README.md
