--- plugins/template/CHANGELOG.md
@@ -1,12 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
-adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
-## 1.0.0 (2018-07-24)
-
-### Added
-
-- Created plugin template and initial documentation
