--- plugins/template/examples/example.py
@@ -1,20 +1,0 @@
-import cmd2_myplugin
-
-import cmd2
-
-
-class Example(cmd2_myplugin.MyPlugin, cmd2.Cmd):
-    """An class to show how to use a plugin."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        # gotta have this or neither the plugin or cmd2 will initialize
-        super().__init__(*args, **kwargs)
-
-    @cmd2_myplugin.empty_decorator
-    def do_something(self, _arg) -> None:
-        self.poutput('this is the something command')
-
-
-if __name__ == '__main__':
-    app = Example()
-    app.cmdloop()
