--- plugins/template/noxfile.py
@@ -1,7 +1,0 @@
-import nox
-
-
-@nox.session(python=['3.10', '3.11', '3.12', '3.13', '3.14'])
-def tests(session) -> None:
-    session.install('invoke', './[test]')
-    session.run('invoke', 'pytest', '--junit', '--no-pty')
