--- plugins/template/tests/__init__.py
@@ -1,2 +1,0 @@
-#
-# empty file to create a package
